backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-cors==5.0.1
flask-bcrypt==1.0.1
flask-migrate==4.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
